refactor(console): split component and non-component exports apart
react-refresh/only-export-components fired on three files that mixed a
component with something else, which costs Fast Refresh on every edit to
them. Each is split along the line the rule draws:

- rjsf-templates.tsx keeps only the template registry; IconButton and
  ArrayFieldItemTemplate move to their own modules.
- command-palette-context.ts now owns the context and useCommandPalette;
  command-palette-provider.tsx keeps only the provider component.
- tenant-context.tsx now owns the context, useTenantContext and
  tenantDisplayName; TenantProvider moves to tenant-provider.tsx.

tenant-context.tsx keeps its path because ~25 modules import
useTenantContext from it — only the four TenantProvider importers move.
The provider's private displayName helper is dropped for the identical
exported tenantDisplayName. Pure moves otherwise: no logic changed.
