+++ packages/system/dashboard/images/console/apps/console/src/components/command-palette/command-palette-provider.tsx
@@ -1,15 +1,6 @@
-import { createContext, useContext, useState, useCallback, useMemo } from "react"
+import { useState, useCallback, useMemo } from "react"
 import { useHotkey } from "./use-hotkey"
-
-interface CommandPaletteContextValue {
-  open: boolean
-  setOpen: (open: boolean) => void
-  toggle: () => void
-}
-
-const CommandPaletteContext = createContext<CommandPaletteContextValue | undefined>(
-  undefined
-)
+import { CommandPaletteContext } from "./command-palette-context.ts"
 
 export function CommandPaletteProvider({ children }: { children: React.ReactNode }) {
   const [open, setOpen] = useState(false)
@@ -30,12 +21,4 @@ export function CommandPaletteProvider({ children }: { children: React.ReactNode
       {children}
     </CommandPaletteContext.Provider>
   )
-}
-
-export function useCommandPalette() {
-  const context = useContext(CommandPaletteContext)
-  if (!context) {
-    throw new Error("useCommandPalette must be used within CommandPaletteProvider")
-  }
-  return context
 }
