+++ packages/system/dashboard/images/console/apps/console/src/lib/tenant-context.tsx
@@ -1,16 +1,8 @@
-import {
-  createContext,
-  useContext,
-  useEffect,
-  useMemo,
-  useState,
-  type ReactNode,
-} from "react"
-import { useK8sList } from "@cozystack/k8s-client"
+import { createContext, useContext } from "react"
 import type { TenantNamespace } from "@cozystack/types"
-import { SELECTED_TENANT_KEY, TENANT_NAMESPACE_PREFIX } from "./constants.ts"
+import { TENANT_NAMESPACE_PREFIX } from "./constants.ts"
 
-interface TenantContextValue {
+export interface TenantContextValue {
   /**
    * Flat list of every TenantNamespace in the cluster, ordered by display
    * name (the namespace prefix stripped). Callers typically only need
@@ -27,70 +19,9 @@ interface TenantContextValue {
   error: unknown
 }
 
-const TenantContext = createContext<TenantContextValue | null>(null)
-
-function displayName(ns: TenantNamespace): string {
-  const name = ns.metadata.name
-  return name.startsWith(TENANT_NAMESPACE_PREFIX)
-    ? name.slice(TENANT_NAMESPACE_PREFIX.length)
-    : name
-}
-
-export function TenantProvider({ children }: { children: ReactNode }) {
-  const [selectedTenant, setSelectedTenant] = useState<string | null>(() => {
-    if (typeof window === "undefined") return null
-    return window.localStorage.getItem(SELECTED_TENANT_KEY)
-  })
-
-  const ns = selectedTenant ? `${TENANT_NAMESPACE_PREFIX}${selectedTenant}` : null
-
-  // TenantNamespace is cluster-scoped, filter by parent tenant label
-  // to show only child tenants of the selected tenant
-  const labelSelector = ns ? `tenant.cozystack.io/${ns}` : undefined
-
-  const list = useK8sList<TenantNamespace>(
-    {
-      apiGroup: "core.cozystack.io",
-      apiVersion: "v1alpha1",
-      plural: "tenantnamespaces",
-    },
-    { labelSelector }
-  )
-
-  const tenants = useMemo<TenantNamespace[]>(() => {
-    return (list.data?.items ?? [])
-      .slice()
-      .sort((a, b) => displayName(a).localeCompare(displayName(b)))
-  }, [list.data])
-
-  useEffect(() => {
-    if (!tenants.length) return
-    if (selectedTenant && tenants.some((t) => displayName(t) === selectedTenant)) return
-    const fallback =
-      tenants.find((t) => displayName(t) === "root") ?? tenants[0]
-    setSelectedTenant(displayName(fallback))
-  }, [tenants, selectedTenant])
-
-  const selectTenant = (name: string) => {
-    setSelectedTenant(name)
-    try {
-      window.localStorage.setItem(SELECTED_TENANT_KEY, name)
-    } catch {
-      // ignore storage quota / private-mode failures
-    }
-  }
-
-  const value: TenantContextValue = {
-    tenants,
-    selectedTenant,
-    selectTenant,
-    tenantNamespace: list.isLoading ? null : ns,
-    isLoading: list.isLoading,
-    error: list.error,
-  }
-
-  return <TenantContext.Provider value={value}>{children}</TenantContext.Provider>
-}
+// TenantProvider lives in tenant-provider.tsx: a file exporting a component
+// alongside a context or a hook breaks Fast Refresh.
+export const TenantContext = createContext<TenantContextValue | null>(null)
 
 export function useTenantContext(): TenantContextValue {
   const ctx = useContext(TenantContext)
@@ -102,5 +33,8 @@ export function useTenantContext(): TenantContextValue {
  * Pull the display name of a TenantNamespace (no `tenant-` prefix).
  */
 export function tenantDisplayName(ns: TenantNamespace): string {
-  return displayName(ns)
+  const name = ns.metadata.name
+  return name.startsWith(TENANT_NAMESPACE_PREFIX)
+    ? name.slice(TENANT_NAMESPACE_PREFIX.length)
+    : name
 }
